Bug: Unable to use ActionShot in Capybara tests with phantomjs driver
ActionShot was using the availability of window.callPhantom to work out
whether it was in capture mode or not. callPhantom is available when
using phantomjs to run Capybara tests so ActionShot was incorrectly
going into "capture" mode. This has been fixed with a simple callback.
ActionShot returns true to the call callPhantom("actionshot:init").